Rework GM damage editor: paperdoll, GM role authority, capital ships
Checkpoint of the damage-editor branch work.

GM role authority
- Only the Game Master edits damage when a GM is present; without one,
  players edit their own units (LobbyActions.canEditDamage).
- The host's "Allow Game Master" game option is now the single toggle.
  Removed the per-client Client Settings checkbox and its preference.
  Disabling the option strips any active GM (TWGameManager) and gates the
  in-game GM tools (MapMenu) and the lobby button (ChatLounge).
- The lobby Game Master button drops its redundant check icon; the toggle
  already shows one.
- The window title marks the local player with " (GM)" (ClientGUI).

Damage editor dialog
- Split UnitEditorDialog into DamageEditorDiagram (armor diagram, split
  pane, coloring, pending preview) and PreExistingDamageRoller (roll,
  snapshots, crit assignment). Dialog drops from ~940 to ~460 lines.
- Renamed the gameMaster flag to offerGameMasterTools.
- Split the pre-existing "Roll and Apply" button into Roll and Apply.
- Added a scaled minimum dialog size.

Paperdoll
- The unit display armor diagram now stripes crit locations and scales to
  fill its panel, with crisp label text at scale (PicMap draws labels at
  native resolution). The diagram sits top-center.

Capital ships
- Weapons and ammo get a panel for their own location instead of piling
  into the general panel; the general panel keeps ship systems and
  transport bays. Bay member weapons are hidden, since the bay is the crit.
- The paperdoll's SI section selects the hull's panel, which holds the
  Structural Integrity and the hull's own equipment.
- Panels wrap into columns of 15 rows for large units like warships.
- Capital fighters draw their armor grid even when capital armor was never
  set (CapitalFighterMapSet).

Sprite damage decal
- The lobby now refreshes on entity change, and the board view reloads a
  unit's image when the change carries no old state, so the damage decal
  keeps up with edits.
