+++ megamek/resources/megamek/client/messages.properties
@@ -724,7 +724,7 @@ ChatLounge.butPrintList.printing=Loading print dialog
 ChatLounge.butShrink=<
 ChatLounge.butSkills=Random Skills...
 ChatLounge.butGameMaster=GM Mode
-ChatLounge.editDamage.notAllowed.tooltip=You may only edit the damage of your own units. The Game Master may edit the damage of any unit.
+ChatLounge.editDamage.notAllowed.tooltip=The Game Master edits the damage of every unit. When there is no Game Master, you may edit the damage of your own units.
 ChatLounge.butGameMaster.held=GM Mode - {0}
 ChatLounge.butGameMaster.tooltip.notAllowed=This game does not allow a Game Master. The host can turn it on in the game options.
 ChatLounge.butGameMaster.tooltip=Take the Game Master role, which allows editing damage, equipment and the board in play.
@@ -969,6 +969,7 @@ CFRAPDSAssign.Message=Pick which attack the APDS on {0} will target.\n
 ClientGUI.ChatWindow=Accessibility Window
 ClientGUI.BotCommand=Bot Command
 ClientGUI.clientTitleSuffix=\ - MegaMek
+ClientGUI.titleGameMaster=\ (GM)
 ClientGUI.titleRoundPhase=\ - Round {0} - {1} phase
 ClientGUI.dialogMovementReport=Movement Report
 ClientGUI.dialogTacticalGeniusReport=Tactical Genius Report
@@ -1512,8 +1513,6 @@ CommonSettingsDialog.sensorMaxDrawRange=Max Range to Draw Sensor / Visual Rings
 CommonSettingsDialog.sensorMaxDrawRange.tooltip=Large values will slow down game, can set to 0 to turn off completely
 CommonSettingsDialog.showArmorMiniVisTT=Show mini armor visualization in the tooltip.
 CommonSettingsDialog.showDamageDecal=Show damage to units on the unit icon
-CommonSettingsDialog.allowGameMasterMode=Allow GM mode
-CommonSettingsDialog.allowGameMasterMode.tooltip=Offer the Game Master role and its tools. Unchecked, the lobby has no Game Master button and this client cannot take the role.
 CommonSettingsDialog.showDamageLevel=Show damage to units on the unit label
 CommonSettingsDialog.showIPAddressesInChat.tooltip=<html>If enabled, all server and client IP addresses will be logged in the public chat box.<br>Enabling this can allow information disclosure, such as private IP addresses of the server.</html>
 CommonSettingsDialog.showIPAddressesInChat=Show IP Addresses in Chat (WARNING: Security Sensitive)
@@ -4546,10 +4545,13 @@ UnitEditorDialog.menLeft=Men Left
 UnitEditorDialog.motiveDamage=Motive Damage
 UnitEditorDialog.points=Points
 UnitEditorDialog.preExistingDamage=Pre-Existing Damage
-UnitEditorDialog.preExistingDamage.apply=Roll and Apply
+UnitEditorDialog.preExistingDamage.roll=Roll
+UnitEditorDialog.preExistingDamage.roll.tooltip=Rolls random pre-existing damage per the First Succession War rules (FSW p.144) and shows it on the armor diagram. Roll again to reroll, or Restore Unit to return it to full health.
+UnitEditorDialog.preExistingDamage.apply=Apply
+UnitEditorDialog.preExistingDamage.apply.tooltip=Sets the shown damage on the unit and closes the editor.
 UnitEditorDialog.preExistingDamage.reset=Restore Unit
 UnitEditorDialog.preExistingDamage.reset.tooltip=Restores the unit to fully repaired: armor and structure to maximum and all critical hits cleared. Press Okay to keep it.
-UnitEditorDialog.preExistingDamage.tooltip=Rolls random pre-existing damage per the First Succession War rules (FSW p.144). The result fills in the controls below; press Okay to keep it, Roll and Apply again to reroll, or Restore Unit to return it to full health.
+UnitEditorDialog.preExistingDamage.tooltip=Rolls random pre-existing damage per the First Succession War rules (FSW p.144). Pick a level, then Roll to preview it on the armor diagram and Apply to set it on the unit.
 UnitEditorDialog.preExistingDamage.NONE=None
 UnitEditorDialog.preExistingDamage.LIGHT=Light (1 point per 5 tons, no crits)
 UnitEditorDialog.preExistingDamage.MODERATE=Moderate (2 points per 5 tons, 1 crit)
